refactor: REST ful에 맞게 변경

+++ banner/views.py
@@ -94,16 +94,6 @@ def create_banner(request):
         404: '배너 조회 실패',
     }
 )
-@api_view(['GET'])
-def get_banner(request, banner_id):
-    try:
-        banner = Banner.objects.get(id=banner_id)
-        serializer = BannerDetailSerializer(banner)
-        return Response({"code": 200, "message": "배너 조회 성공", "data": serializer.data}, status=status.HTTP_200_OK)
-    except Banner.DoesNotExist:
-        return Response({"code": 404, "message": "배너 조회 실패"}, status=status.HTTP_404_NOT_FOUND)
-
-
 @swagger_auto_schema(
     method='put',
     operation_id='배너 수정',
@@ -114,36 +104,6 @@ def get_banner(request, banner_id):
         404: '배너 수정 실패',
     }
 )
-@api_view(['PUT'])
-def update_banner(request, banner_id):
-    try:
-        banner = Banner.objects.get(id=banner_id)
-    except Banner.DoesNotExist:
-        return Response({"code": 404, "message": "배너 수정 실패"}, status=status.HTTP_404_NOT_FOUND)
-
-    serializer = BannerUpdateSerializer(banner, data=request.data)
-    if serializer.is_valid():
-        item_name = serializer.validated_data.get('item_name')
-        item_concept = serializer.validated_data.get('item_concept')
-
-        ad_text = generate_ad_text(item_name, item_concept)
-
-        banner.item_name = item_name
-        banner.item_concept = item_concept
-        banner.ad_text = ad_text
-        banner.save()
-
-        response_data = {
-            "code": 200,
-            "message": "배너 수정 성공",
-            "data": BannerDetailSerializer(banner).data,
-        }
-
-        return Response(response_data, status=status.HTTP_200_OK)
-
-    return Response({"code": 400, "message": "배너 수정 실패", "errors": serializer.errors},
-                    status=status.HTTP_400_BAD_REQUEST)
-
 @swagger_auto_schema(
     method='delete',
     operation_id='배너 삭제',
@@ -153,11 +113,43 @@ def update_banner(request, banner_id):
         404: '배너 삭제 실패',
     }
 )
-@api_view(['DELETE'])
-def delete_banner(request, banner_id):
+@api_view(['GET', 'PUT', 'DELETE'])
+def handle_banner(request, banner_id):
     try:
         banner = Banner.objects.get(id=banner_id)
+    except Banner.DoesNotExist:
+        return Response({"code": 404, "message": "배너 조회 실패"}, status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET':
+        serializer = BannerDetailSerializer(banner)
+        return Response({"code": 200, "message": "배너 조회 성공", "data": serializer.data}, status=status.HTTP_200_OK)
+
+    elif request.method == 'PUT':
+        serializer = BannerUpdateSerializer(banner, data=request.data)
+        if serializer.is_valid():
+            item_name = serializer.validated_data.get('item_name')
+            item_concept = serializer.validated_data.get('item_concept')
+
+            ad_text = generate_ad_text(item_name, item_concept)
+
+            banner.item_name = item_name
+            banner.item_concept = item_concept
+            banner.ad_text = ad_text
+            banner.save()
+
+            response_data = {
+                "code": 200,
+                "message": "배너 수정 성공",
+                "data": BannerDetailSerializer(banner).data,
+            }
+
+            return Response(response_data, status=status.HTTP_200_OK)
+
+        return Response({"code": 400, "message": "배너 수정 실패", "errors": serializer.errors},
+                        status=status.HTTP_400_BAD_REQUEST)
+
+    elif request.method == 'DELETE':
         banner.delete()
         return Response({"code": 200, "message": "배너 삭제 성공"}, status=status.HTTP_200_OK)
-    except Banner.DoesNotExist:
-        return Response({"code": 404, "message": "배너 삭제 실패"}, status=status.HTTP_404_NOT_FOUND)
+
+    return Response({"code": 405, "message": "Method not allowed"}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
